+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 10.1
 - Filter illegal chars in ingredient search
+- Added orientation when exporting to a formula for selling
 ### Version 10.0
 - Discord server link added
 - Added archive option when deleting a formula